move to own declaration files
